Fix error if invalid "content-type" is passed via querystring
If an invalid "content-type" is passed via the querystring,
the following error will be thrown by Express:
> TypeError: invalid media type

The fix is to use `content-type` to validate the querystring.
Respond with 500 and error message if content-type is invalid.

Refactor the header(s) set/override logic so if an error is
caught, the response can be sent without causing the following:
> Error: Can't set headers after they are sent.